linux: the sources can be incrementally retrieved
This is a rather large commit, because I built the foundations of mktcb
on the fly, a I learnt new rust features and discover interesting
crates.

Currently, this is what works:

- options parsing (with clap)
- file download (with curl-rs) with a progress bar (with indicatif)
- error management (with snafu)
- decompression (calling tar -xf and crate xz2)
- patching files (calling patch)
- robustness to sigint (preventing exit mid-patches)
- TOML config loading (toml and serde)

Local patches are not yet applied, and the 'make' command cannot be run
yet.  U-boot has not been implemented.